添加自定义的异常处理机制；实现回复功能

+++ src/main/java/com/mrh/community/controller/QuestionController.java
@@ -1,9 +1,7 @@
 package com.mrh.community.controller;
 
 import com.mrh.community.dto.QuestionDTO;
-import com.mrh.community.mapper.QuestionMapper;
 import com.mrh.community.service.QuestionService;
-import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -23,7 +21,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id")Integer id,
+    public String question(@PathVariable(name = "id")Long id,
                            Model model)
     {
         QuestionDTO questionDTO = questionService.getById(id);
